перенес planeMeshController в отдельный файл
